refactoring chiller names
